implement click-to-copy in vm_table_headers
