+++ src/client/gui/lib/vm_details/vm_details_general.dart
@@ -12,51 +12,7 @@ import 'vm_action_buttons.dart';
 import 'vm_details.dart';
 import 'vm_status_icon.dart';
 import '../tooltip.dart';
-
-class CopyableText extends StatefulWidget {
-  final String text;
-  final TextStyle? style;
-
-  const CopyableText(this.text, {super.key, this.style});
-
-  @override
-  State<CopyableText> createState() => _CopyableTextState();
-}
-
-class _CopyableTextState extends State<CopyableText> {
-  bool _copied = false;
-
-  void _copyToClipboard() async {
-    await Clipboard.setData(ClipboardData(text: widget.text));
-    setState(() => _copied = true);
-  }
-
-  void _resetCopied() {
-    if (_copied) {
-      setState(() => _copied = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      onExit: (_) => _resetCopied(),
-      child: GestureDetector(
-        onTap: _copyToClipboard,
-        child: Tooltip(
-          message: _copied ? 'Copied' : 'Click to copy',
-          child: Text(
-            widget.text,
-            style: widget.style,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-          ),
-        ),
-      ),
-    );
-  }
-}
+import '../copyable_text.dart';
 
 class VmDetailsHeader extends ConsumerWidget {
   final String name;
